Update Logger to automatically assign tacking id to all error event
