+++ src/utilities/errorHandling/globalErrorHandler.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-console */
 const { v4: uuidv4 } = require('uuid');
 const { CustomError } = require('./customError');
-const Logger = require('../startup/logger');
+const Logger = require('../../startup/logger');
 
 /**
  * Custom error handler middleware for global unhandled errors. Make it the last middleware since it returns a response and do not call next().
-*/
+ */
 function globalErrorHandler(err, req, res, next) {
   /**
    * Notes:
@@ -15,7 +15,7 @@ function globalErrorHandler(err, req, res, next) {
    */
   const trackingId = uuidv4();
   const errorMessage = `An internal error has occurred. If the issue persists, please contact the administrator and provide the trakcing ID: ${trackingId}`;
-  
+
   let transactionName = '';
   const requestData = req.body && req.method ? JSON.stringify(req.body) : null;
 
@@ -27,19 +27,17 @@ function globalErrorHandler(err, req, res, next) {
   }
 
   // transactionName = transactionName.concat(' ', 'Tracking ID: ', eventId);
-  if(!err){
-    transactionName = 'Critical: err parameter is missing. This is probably due to an improper error handling in the code.';
+  if (!err) {
+    transactionName =
+      'Critical: err parameter is missing. This is probably due to an improper error handling in the code.';
   }
   // Log the error to Sentry if not in local environment
   if (process.env.NODE_ENV === 'development' || process.env.NODE_ENV === 'production') {
-    Logger.logException(
-      err,
-      transactionName,
-      requestData,
-      trackingId
-    );
+    Logger.logException(err, transactionName, requestData, trackingId);
   } else {
-    console.log(`An error occurred. Transaction: ${transactionName} \nRequest Data: ${requestData}`);
+    console.log(
+      `An error occurred. Transaction: ${transactionName} \nRequest Data: ${requestData}`,
+    );
     console.error(err);
   }
